+++ src/Snipdeck.App/Views/ShellContentTemplateSelector.cs
@@ -19,7 +19,9 @@ namespace Snipdeck.App.Views
 
         public DataTemplate? TrashTemplate { get; set; }
 
-        public DataTemplate? GlobalParametersTemplate { get; set; }
+        public DataTemplate? SharedParametersTemplate { get; set; }
+
+        public DataTemplate? TagIconsTemplate { get; set; }
 
         protected override DataTemplate? SelectTemplateCore(object item)
         {
@@ -29,7 +31,8 @@ namespace Snipdeck.App.Views
                 CliViewModel => CliTemplate,
                 SettingsViewModel => SettingsTemplate,
                 TrashViewModel => TrashTemplate,
-                GlobalParametersViewModel => GlobalParametersTemplate,
+                SharedParametersViewModel => SharedParametersTemplate,
+                TagIconsViewModel => TagIconsTemplate,
                 _ => null,
             };
         }